# database: Optimise CountUserAddedRepos